removed pkg and moved its subpackages one level up
